Xml Writer fixed

+++ BookMaster/BooksLibrary/BooksWriter.cs
@@ -5,6 +5,8 @@ using System.Data;
 using System.Xml.Serialization;
 using System.Collections;
 using System.Collections.Generic;
+using System.Linq;
+using System.Xml.Linq;
 
 
 namespace BookMaster.BooksLibrary
@@ -15,10 +17,12 @@ namespace BookMaster.BooksLibrary
 
         public static void AddBook(int Isbn, string Author, string Title)
         {
-            // Creates an instance of the XmlSerializer class;  
-            // XmlSerializer serializer = specifies the type of object to serialize.  
-            //XmlSerializer serializer = new XmlSerializer(typeof(PurchaseOrder));
-            XmlSerializer serializer = new XmlSerializer(typeof(BooksArray));
+            if (File.Exists(filePath)) { 
+                // Creates an instance of the XmlSerializer class;  
+                // XmlSerializer serializer = specifies the type of object to serialize.  
+                //XmlSerializer serializer = new XmlSerializer(typeof(PurchaseOrder));
+                XmlSerializer serializer = new XmlSerializer(typeof(BooksArray));
+
             TextWriter writer = new StreamWriter(filePath);
 
             BooksArray booksInst = new BooksArray();
@@ -55,60 +59,41 @@ namespace BookMaster.BooksLibrary
             // serializer.Serialize(writer, booksInst);
 
             writer.Close();
+            }
         }
 
         public static void AddBookNew(int Isbn, string Author, string Title)
         {
-            XmlSerializer serializer = new XmlSerializer(typeof(BooksArray));
-            // A FileStream is needed to read the XML document.  
-            FileStream fs = new FileStream(filePath, FileMode.Open);
-            // Declares an object variable of the type to be deserialized.  
-            BooksArray po;
-            // Uses the Deserialize method to restore the object's state   
-            // with data from the XML document. */  
-            po = (BooksArray)serializer.Deserialize(fs);
-
-
-            // Creates an OrderedItem.  
-            Book i1 = new Book();
-
-
-            i1.Isbn = Isbn;//1;
-            i1.Author = Author; //"Sigmar";
-            i1.Title = Title;//"Cat world";
-
-
-            List<Book> items = po.Books;
-
-
-            ArrayList a1 = new ArrayList();
-
-            a1.Add(i1);
-            a1.Add(items);
-
-            BooksArray booksArr = new BooksArray();
-
-            // Array stuff = a1.ToArray();
-            // booksArr.Books = stuff;
-
-            TextWriter writer = new StreamWriter(filePath);
-
-
-            // Serializes the purchase order, and closes the TextWriter.  
-            serializer.Serialize(writer, items);
-
-            // serializer.Serialize(writer, booksInst);
-
-            writer.Close();
-
-            foreach (Book oi in a1)
+            if (File.Exists(filePath))
             {
-                Console.WriteLine("\t" +
-                oi.Isbn + "\t" +
-                oi.Title + "\t" +
-                oi.Author + "\t");
+                XmlSerializer serializer = new XmlSerializer(typeof(BooksArray));
+                FileStream fs = new FileStream(filePath, FileMode.Open);
+                XmlReader reader = XmlReader.Create(fs);
 
+
+                BooksArray i = (BooksArray)serializer.Deserialize(reader);
+                fs.Close();
+
+                Book i1 = new Book();
+
+
+                i1.Isbn = Isbn;//1;
+                i1.Author = Author; //"Sigmar";
+                i1.Title = Title;//"Cat world";
+
+                i.Books.Add(i1);
+                foreach (Book p in  i.Books)
+                {
+                    Console.WriteLine(p.Isbn);
+                }
+
+                TextWriter writer = new StreamWriter(filePath);
+                serializer.Serialize(writer, i);
+                writer.Close();
             }
+           
+
+
         }
 
         public static void ListBooks()
@@ -140,7 +125,7 @@ namespace BookMaster.BooksLibrary
             */
             // Reads the list of ordered items.  
             List<Book> items = po.Books;
-            Console.WriteLine("Items to be shipped:");
+         
             foreach (Book oi in items)
             {
                 Console.WriteLine("\t" +
@@ -149,6 +134,7 @@ namespace BookMaster.BooksLibrary
                 oi.Author + "\t");
 
             }
+            fs.Close();
         }
     }
 }
